version 3.4

+++ src/mutual.information.c
@@ -49,7 +49,7 @@ SEXP result;
   PROTECT(result = allocVector(REALSXP, 2));
   res = REAL(result);
   res[0] = c_mi(xx, &llx, yy, &lly, &num);
-  res[1] = (llx - 1) * (lly - 1);
+  res[1] = (double)(llx - 1) * (double)(lly - 1);
 
   /* rescale to match the G^2 test. */
   if (isTRUE(gsquare))
@@ -118,7 +118,7 @@ SEXP result;
   PROTECT(result = allocVector(REALSXP, 2));
   res = REAL(result);
   res[0] = c_cmi(xx, &llx, yy, &lly, zz, &llz, &num);
-  res[1] = (llx - 1) * (lly - 1) * llz;
+  res[1] = (double)(llx - 1) * (double)(lly - 1) * (double)llz;
 
   /* rescale to match the G^2 test. */
   if (isTRUE(gsquare))
