+++ src/pearson.x2.c
@@ -13,7 +13,7 @@ SEXP result;
   PROTECT(result = allocVector(REALSXP, 2));
   res = REAL(result);
   res[0] = 0;
-  res[1] = (llx - 1) * (lly - 1);
+  res[1] = (double)(llx - 1) * (double)(lly - 1);
 
   /* initialize the contingency table and the marginal frequencies. */
   n = alloc2dcont(llx, lly);
@@ -65,7 +65,7 @@ SEXP result;
   PROTECT(result = allocVector(REALSXP, 2));
   res = REAL(result);
   res[0] = 0;
-  res[1] = (llx - 1) * (lly - 1) * llz;
+  res[1] = (double)(llx - 1) * (double)(lly - 1) * (double)llz;
 
   /* initialize the contingency table and the marginal frequencies. */
   n = alloc3dcont(llx, lly, llz);
